[feat] Add authorization to admin users (level 9)

+++ schedule-management/src/main/java/com/sparta/schedulemanagement/service/UserService.java
@@ -2,17 +2,23 @@ package com.sparta.schedulemanagement.service;
 
 import com.sparta.schedulemanagement.config.PasswordEncoder;
 import com.sparta.schedulemanagement.dto.LoginRequestDto;
+import com.sparta.schedulemanagement.dto.SignUpRequestDto;
 import com.sparta.schedulemanagement.dto.UserRequestDto;
 import com.sparta.schedulemanagement.dto.UserResponseDto;
 import com.sparta.schedulemanagement.entity.User;
+import com.sparta.schedulemanagement.entity.UserRoleEnum;
 import com.sparta.schedulemanagement.jwt.JwtUtil;
 import com.sparta.schedulemanagement.repository.UserRepository;
+import jakarta.servlet.ServletRequest;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.io.IOException;
+import java.net.http.HttpHeaders;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -31,6 +37,8 @@ public class UserService {
         );
     }
 
+    private final String ADMIN_TOKEN = "AAABnvxRVklrnYxKZ0aHgTBcXukeZygoC";
+
     public List<UserResponseDto> getAllUsers() {
         List<User> users = userRepository.findAll();
         List<UserResponseDto> userResponseDtos = new ArrayList<>();
@@ -40,10 +48,10 @@ public class UserService {
         return userResponseDtos;
     }
 
-    public UserResponseDto signUp(UserRequestDto userRequestDto, HttpServletResponse res) {
-        String username = userRequestDto.getUsername();
-        String email = userRequestDto.getEmail();
-        String password = passwordEncoder.encode(userRequestDto.getPassword());
+    public UserResponseDto signUp(SignUpRequestDto signUpRequestDto, HttpServletResponse res) {
+        String username = signUpRequestDto.getUsername();
+        String email = signUpRequestDto.getEmail();
+        String password = passwordEncoder.encode(signUpRequestDto.getPassword());
 
         Optional<User> checkUsername = userRepository.findByUsername(username);
         if(checkUsername.isPresent()) {
@@ -54,11 +62,20 @@ public class UserService {
             throw new IllegalArgumentException("User with email " + email + " already exists");
         }
 
-        String token = jwtUtil.createToken(username);
+        UserRoleEnum role = UserRoleEnum.USER;
+
+        if(signUpRequestDto.isAdmin()) {
+            if(!ADMIN_TOKEN.equals(signUpRequestDto.getAdminToken())) {
+                throw new IllegalArgumentException("Admin token does not match");
+            }
+            role = UserRoleEnum.ADMIN;
+        }
+
+        String token = jwtUtil.createToken(username, role);
         jwtUtil.addJwtToCookie(token, res);
 
-        User user = new User(userRequestDto);
-        user.setEncodedPassword(password);
+        User user = new User(username, password, email, role);
+
         userRepository.save(user);
         return new UserResponseDto(user);
     }
@@ -68,15 +85,27 @@ public class UserService {
     }
 
     @Transactional
-    public UserResponseDto updateUser(int id, @Valid UserRequestDto userRequestDto) {
+    public UserResponseDto updateUser(int id, @Valid UserRequestDto userRequestDto, HttpServletResponse res, ServletRequest req) throws IOException {
         User user = findUserById(id);
-        user.update(userRequestDto);
-        return new UserResponseDto(user);
+
+        UserRoleEnum currentRole = (UserRoleEnum) req.getAttribute("user_role");
+        if(currentRole == UserRoleEnum.ADMIN) {
+            user.update(userRequestDto);
+            return new UserResponseDto(user);
+        }
+
+        res.sendError(HttpServletResponse.SC_FORBIDDEN, "Only ADMIN users can update user");
+        return null;
     }
 
-    public String deleteUser(int id) {
-        userRepository.deleteById(id);
-        return "User with id " + id + " deleted";
+    public String deleteUser(int id, HttpServletResponse res, ServletRequest req) throws IOException {
+        UserRoleEnum currentRole = (UserRoleEnum) req.getAttribute("user_role");
+        if(currentRole == UserRoleEnum.ADMIN) {
+            userRepository.deleteById(id);
+            return "User with id " + id + " deleted";
+        }
+        res.sendError(HttpServletResponse.SC_FORBIDDEN, "Only ADMIN users can delete user");
+        return "Only ADMIN users can delete user";
     }
 
     public String login(LoginRequestDto loginRequestDto, HttpServletResponse res) {
@@ -92,7 +121,7 @@ public class UserService {
             return "Invalid email or password";
         }
 
-        String token = jwtUtil.createToken(user.get().getUsername());
+        String token = jwtUtil.createToken(user.get().getUsername(), user.get().getRole());
         jwtUtil.addJwtToCookie(token, res);
         return token;
     }
